chore: Add: - add_actor - play_status as well as some refactoring with interfaces

+++ src/interfaces.rs
@@ -2,6 +2,7 @@
 // use binary_utils::{BinaryStream, IBinaryStream, IBufferWrite};
 
 /// A 12 byte value prefixed with 3 floats, x y and z.
+#[derive(Copy, Clone)]
 pub struct Vector3 {
      pub x: f32,
      pub y: f32,
@@ -9,19 +10,22 @@ pub struct Vector3 {
 }
 
 /// A 8 byte value prefixed with 2 floats, x, z.
+#[derive(Copy, Clone)]
 pub struct Vector2 {
      pub x: f32,
      pub z: f32
 }
 
 /// A 3 - 15 byte struct
+#[derive(Copy, Clone)]
 pub struct Coordinates {
-     pub x: isize,
-     pub y: usize,
-     pub z: isize
+     pub x: i64,
+     pub y: i64,
+     pub z: i64
 }
 
 /// 15 byte structure for location
+#[derive(Copy, Clone)]
 pub struct Location {
      pub x: f32,
      pub y: f32,
